// Frontend/src/api/models/TaskResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Task response schema
 */
export type TaskResponse = {
    name: string;
    description?: (string | null);
    start_date?: (string | null);
    end_date?: (string | null);
    owner_id?: (number | null);
    id: number;
    investment_id: number;
    status?: (string | null);
    created_at?: (string | null);
};