// Frontend/src/api/models/TaskCreate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Task creation schema
 */
export type TaskCreate = {
    name: string;
    description?: (string | null);
    start_date?: (string | null);
    end_date?: (string | null);
    owner_id?: (number | null);
};